requirements.txt
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.6
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
commonmark==0.9.1
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.18
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
pyarrow==6.0.1
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.6.2
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.12.0
pylint==2.13.9
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==1.1.6
pyzmq==23.2.0
requests==2.27.1
rich==12.4.4
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
streamlit==1.10.0
terminado==0.12.1
testpath==0.6.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.5.4
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wrapt==1.14.1
zipp==3.6.0
